requirements.txt
torch
torchvision
scikit-learn
numpy
pillow